redis_image_loader/requirements.txt
django
redis
Pillow
